demo/JcsPicDemo/requirements.txt
baidu-aip==2.2.18.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
jcs-sdk==1.7.25
Jinja2==3.0.1
logzero==1.7.0
MarkupSafe==2.0.1
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
